Refactor player stats and blade upgrade system
Статы игрока теперь используют базовые значения и методы пересчета для максимального здоровья, Рейки, уклонения, точности и урона. Эффекты благословения и улучшения статов обновляют базовые статы и вызывают перерасчет. Меню апгрейдов клинка теперь поддерживает апгрейды шанса крита и способности духа. Боевая логика использует пересчитанные статы игрока для расчета хитов, урона и критов.  Логика сохранения/загрузки обновлена для сохранения базовых статов.

+++ Blade&Kami/combat_logic.h
@@ -4,8 +4,8 @@
 #include <algorithm>
 
 #include "enemy.h"
-#include "player.h"
 #include "game_state.h"
+#include "player.h"
 
 class CombatLogic {
  public:
@@ -13,7 +13,7 @@ class CombatLogic {
   static void ProcessEnemyAction(Player& player, Enemy& enemy);
 
   static double CalculatePurificationChance(Player& player, Enemy& enemy);
-  static bool CalculateHit(float attacker_accuracy, float target_evasion);
+  static bool CalculateHit(double attacker_accuracy, double target_evasion);
   static int CalculateDamage(int min_dmg, int max_dmg);
 
   static void OnEnemyKilled(Player& player, Enemy& enemy, GameState& state);
